requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
requests==2.28.0
scipy==1.8.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
